requirements.txt
streamlit==1.35.0
tensorflow==2.16.1
numpy==1.24.0
pillow
opencv-python==4.10.0.82
opencv-python-headless
